requirements.txt
# Core
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.3.4
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Web Intelligence
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# Database (for production)
sqlalchemy==2.0.30
alembic==1.13.1
redis==5.0.4

# AI (when ready to add)
anthropic==0.34.2
openai==1.35.3

# Utils
python-dateutil==2.9.0
